+++ src/components/common/AnnouncementBar/index.js
